moved load_audio_task from fad.py to utils.py
