+++ frechet_audio_distance/utils.py
@@ -1,0 +1,25 @@
+import resampy
+import numpy as np
+import soundfile as sf
+
+
+def load_audio_task(fname, sample_rate, dtype="float32"):
+    # print("LOAD AUDIO TASK")
+    if dtype not in ['float64', 'float32', 'int32', 'int16']:
+        raise ValueError(f"dtype not supported: {dtype}")
+
+    wav_data, sr = sf.read(fname, dtype=dtype)
+    # For integer type PCM input, convert to [-1.0, +1.0]
+    if dtype == 'int16':
+        wav_data = wav_data / 32768.0
+    elif dtype == 'int32':
+        wav_data = wav_data / float(2**31)
+
+    # Convert to mono
+    if len(wav_data.shape) > 1:
+        wav_data = np.mean(wav_data, axis=1)
+
+    if sr != sample_rate:
+        wav_data = resampy.resample(wav_data, sr, sample_rate)
+
+    return wav_data
